Move functions into  folder and tag controllers
